Convert component TS
